Rename runners-api-slice, invalidate cache, and redirect to created org page
