+++ client/src/App.tsx
@@ -31,7 +31,6 @@ function App({ pca }: AppProps) {
         <div>
           <Routes>
             {Object.values(routes).map((route, index) => {
-              console.log(index);
               if (route.requiresLogin) {
                 return (
                   <Route
