+++ client/src/pages/organizations/createOrganization.tsx
@@ -1,7 +1,7 @@
 import React, { ChangeEvent, useState } from "react";
 import { useNavigate } from "react-router";
 import { Button, Divider, Form, Header, Message } from "semantic-ui-react";
-import { useCreateOrganizationMutation } from "../../slices/runners/runners-api-slice";
+import { useCreateOrganizationMutation } from "../../slices/runners/raceresults-api-slice";
 import BasePage from "../../utils/basePage";
 import routes from "../../utils/routes";
 
@@ -27,7 +27,7 @@ const CreateOrganizationPage = () => {
     createOrganization({ name: name })
       .unwrap()
       .then((createdOrg) =>
-        navigate(routes.organizations.createPath(createdOrg.id))
+        navigate(routes.organization.createPath(createdOrg.id))
       )
       .catch(() => setError(true));
   };
